feat(api): refactor crud operations to new file
